Updated Vulkan to 1.3, including VMA. Ensured new layers and extensions that I wanted for the device, like dynamic rendering, were supported. They seem to be, and I hacked together a pipeline created without a render pass, which is very promising.
Added CommandBuffer-related handles. They're kind of dumb, but the general idea is that the device has resources it has direct ownership over (outside of images and buffers), like the command pool. The command pool then has ownership over command buffers. The kind of initial pass of how to get the device to the command buffer as well as the command pool is kind of sloppy. I started to move away from this implementation, but I want to commit something before I overhaul things to make handles make more sense