Add the internal concept of Handles.
Before, handles were just an int that was type safe, but with no other behavior to it. As I was implementing command buffer/pool handle functionality, a question came up. How does the device internally have some sort of reference to all of the resources it owns and how do the resources themselves have a reference to the device? Some resources, like command buffers, need to know what pool they belong to as well. I expect something similar for descriptors

To solve this, handles were introduced to try and fix this issue. Pools of handles exist now within a device and when a resource is created, the respective handle is set up and given to the caller of the create function. Each handle knows what device it belongs to and if there's a pool, it also has data associated with that as well. Internally, it has an index for its own data as well as a generation counter, which will be used to check validity within the handle system (not currently implemented)

Eventually, it might be good to condense handle pools into one pool of handles, since we're allowing resizing. However, this may restrict the user to only allow for resizing even when they initialize the API to not have resizing. Something to think about