Add Buffer and Queue handles, along with CommandBuffer commands.
Buffers are able to be created and destroyed, which allows for replacing the vertex and index and staging buffers within the sandbox. There is no querying of data about the buffer yet. That will come when I add image and image view creation.

Queues are created at device creation time, so this poses a problem. Currently, there are only one queue that gets allocated between compute and graphics, which is ok but is confusing to the user, since they do not do custom queue set up. This will eventually change, but is fine for now