API at parity with original prototype
Apparition now handles Dynamic Rendering as well as some backbuffer functionality.

Noteable absences from the API are Descriptor Set handling (need to look into bindless to also support that) and Pipeline creation. Pipeline creation poses a possible problem in that it's complex and I want to support the new graphics pipeline library functionality. There may be a way to simplify this, but I will tackle that later on